+++ cpp/content_formatter.hpp
@@ -6,11 +6,11 @@ namespace text {
     class ContentFormatter {
     public:
         virtual ~ContentFormatter() {}
-        virtual std::string format_matrix(double** matrix, const int &size) = 0;
+        virtual std::string format_matrix(long double** matrix, const int &size) = 0;
     };
 
     class ContentFormatterImpl : public ContentFormatter {
     public:
-        std::string format_matrix(double** matrix, const int &size) override;
+        std::string format_matrix(long double** matrix, const int &size) override;
     };
 }
