+++ cpp/memory_manager.cpp
@@ -1,6 +1,6 @@
 #include "memory_manager.hpp"
 
-void memory::MemoryManagerImpl::cleanup_matrix(double** matrix, const int &size) {
+void memory::MemoryManagerImpl::cleanup_matrix(long double** matrix, const int &size) {
     for (int i = 0; i < size; ++i) {
         delete[] matrix[i];
     }
